+++ .eslintrc.json
@@ -1,0 +1,46 @@
+{
+  "extends": [
+    "plugin:@typescript-eslint/recommended",
+    "react-app",
+    "react-app/jest",
+    "naver",
+    "prettier"
+  ],
+  "parser": "@typescript-eslint/parser",
+  "plugins": ["@typescript-eslint"],
+  "rules": {
+    "no-console": 1,
+    "react-hooks/exhaustive-deps": "off",
+    "no-unused-expression": "off",
+    "no-unused-vars": "off",
+    "react/prop-types": "off",
+    "consistent-return": "off",
+    "no-shadow": "off",
+    "spaced-comment": "off",
+    "import/prefer-default-export": "off",
+    "no-alert": "off",
+    "@typescript-eslint/no-empty-function": "off",
+    "import/order": [
+      "error",
+      {
+        "groups": [
+          "external",
+          "builtin",
+          "internal",
+          "parent",
+          "sibling",
+          "index",
+          "object"
+        ],
+        "pathGroups": [
+          {
+            "pattern": "react*",
+            "group": "external",
+            "position": "before"
+          }
+        ],
+        "pathGroupsExcludedImportTypes": ["react*"]
+      }
+    ]
+  }
+}
