+++ docs/기능 목록.md
@@ -6,9 +6,10 @@
 
 ### 설정
 
-- [] prettier
+- [x] prettier
 - [] eslint
 - [] husky
+- [] stylelint
 
 ### 기능
 
@@ -38,9 +39,10 @@
 
 #### 설정
 
-- [] prettier
+- [x] prettier
 - [] eslint
 - [] husky
+- [] stylelint
 
 #### 기능 구현
