+++ src/components/Button/Button.tsx
@@ -1,6 +1,6 @@
-import styled from "styled-components";
-import { DefaultButton } from "./Button.style";
 import { ReactNode } from "react";
+import styled from "styled-components";
+import DefaultButton from "./Button.style";
 
 export default function Button({ children }: { children: ReactNode }) {
   return <StyledButton>{children}</StyledButton>;
