+++ src/components/Button/SubmitButton/SubmitButton.tsx
@@ -1,6 +1,6 @@
-import styled from "styled-components";
-import { DefaultButton } from "../Button.style";
 import { ReactNode } from "react";
+import styled from "styled-components";
+import DefaultButton from "../Button.style";
 
 interface Props {
   isActive: boolean;
