+++ src/components/Button/index.ts
@@ -3,8 +3,10 @@ import ButtonMain from "./Button";
 import SubmitButton from "./SubmitButton";
 import LikeButton from "./LikeButton";
 
-export const Button = Object.assign(ButtonMain, {
+const Button = Object.assign(ButtonMain, {
   Link: LinkButton,
   Submit: SubmitButton,
   Like: LikeButton,
 });
+
+export default Button;
