+++ src/components/Input/FormInput/index.ts
@@ -6,9 +6,11 @@ import {
 } from "./FormInputs";
 import ImageInput from "./ImageInput";
 
-export const FormInput = Object.assign(FormInputMain, {
+const FormInput = Object.assign(FormInputMain, {
   Number: NumberInput,
   Textarea: TextareaInput,
   Tag: TagInput,
   Image: ImageInput,
 });
+
+export default FormInput;
