+++ src/components/Input/index.ts
@@ -1,9 +1,11 @@
-import { FormInput } from "./FormInput";
+import FormInput from "./FormInput";
 import SearchInput from "./SearchInput";
 import SelectInput from "./SelectInput";
 
-export const Input = {
+const Input = {
   Search: SearchInput,
   Select: SelectInput,
   Form: FormInput,
 };
+
+export default Input;
