+++ src/components/Tag/index.ts
@@ -1,5 +1,6 @@
 import TagMain from "./Tag";
 import ProductTag from "./ProductTag";
+
 export { default as TagList } from "./TagList";
 
 export const Tag = Object.assign(TagMain, {
