+++ src/contexts/MarketMain/OrderContext.tsx
@@ -1,4 +1,3 @@
-import { Order } from "@/models/order";
 import {
   Dispatch,
   ReactNode,
@@ -7,6 +6,7 @@ import {
   useContext,
   useState,
 } from "react";
+import { Order } from "models/order";
 
 interface OrderContextInterface {
   order: Order;
@@ -30,18 +30,21 @@ export function OrderProvider({ children }: { children: ReactNode }) {
 
 export function useOrder() {
   const context = useContext(OrderContext);
+
   if (!context) throw new Error("Provider 안에서 사용해야 합니다");
   return context.order;
 }
 
 export function useSetOrder() {
   const context = useContext(OrderContext);
+
   if (!context) throw new Error("Provider 안에서 사용해야 합니다");
   return context.setOrder;
 }
 
 export function useOrderState(): [Order, Dispatch<SetStateAction<Order>>] {
   const context = useContext(OrderContext);
+
   if (!context) throw new Error("Provider 안에서 사용해야 합니다");
   return [context.order, context.setOrder];
 }
