+++ src/contexts/MarketMain/index.tsx
@@ -1,6 +1,7 @@
 import { ReactNode } from "react";
 import { OrderProvider } from "./OrderContext";
 import { PageProvider } from "./PageContext";
+
 export {
   useCurrentPage,
   useSetCurrentPage,
@@ -9,6 +10,7 @@ export {
   useSetTotalPages,
   useTotalPagesState,
 } from "./PageContext";
+
 export { useOrder, useSetOrder, useOrderState } from "./OrderContext";
 
 export default function MarketMainProvider({
