+++ src/index.tsx
@@ -1,9 +1,9 @@
 import React from "react";
+import { BrowserRouter } from "react-router-dom";
 import ReactDOM from "react-dom/client";
 import "styles/index.scss";
 import App from "App";
 import reportWebVitals from "reportWebVitals";
-import { BrowserRouter } from "react-router-dom";
 
 const root = ReactDOM.createRoot(
   document.getElementById("root") as HTMLElement,
