+++ tsconfig.json
@@ -17,7 +17,16 @@
     "jsx": "react-jsx",
     "baseUrl": "src",
     "paths": {
-      "@/*": ["*"]
+      "pages/*": ["pages/*"],
+      "templates/*": ["templates/*"],
+      "components/*": ["components/*"],
+      "styles/*": ["styles/*"],
+      "assets/*": ["assets/*"],
+      "utils/*": ["utils/*"],
+      "hooks/*": ["hooks/*"],
+      "contexts/*": ["contexts/*"],
+      "api/*": ["api/*"],
+      "models/*": ["models/*"]
     }
   },
   "include": ["src"]
